Refactor the class to partial classes to split the implementation of interfaces VsProject and ProjectProperties
Add doc comments
