+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Properties/VsLangProjectProperties.cs
@@ -2,14 +2,16 @@
 
 using System;
 using System.Threading.Tasks;
-using EnvDTE;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
 using VSLangProj;
 using VSLangProj110;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
 {
-    internal class VsLangProjectProperties : VSProject, VSLangProj.ProjectProperties
+    /// <summary>
+    /// See <see cref="VsLangProjectPropertiesProvider"/> for more info.
+    /// </summary>
+    internal partial class VsLangProjectProperties : VSLangProj.ProjectProperties
     {
         private readonly VSProject _vsProject;
         private readonly IUnconfiguredProjectVsServices _projectVsServices;
@@ -25,67 +27,32 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
             _projectProperties = projectProperties;
         }
 
-
-        // Implementation of VSProject to redirect the call to the actual VSProject object
-        public ProjectItem CreateWebReferencesFolder()
+        private void RunFuncTaskSynchronously<T>(Func<T, Task> asyncActionT, T value)
         {
-            return _vsProject.CreateWebReferencesFolder();
+            _projectVsServices.ThreadingService.JoinableTaskFactory.Run(
+                async () =>
+                {
+                    await asyncActionT(value).ConfigureAwait(false);
+                });
         }
 
-        public ProjectItem AddWebReference(string bstrUrl)
+        public prjOutputTypeEx OutputTypeEx
         {
-            return _vsProject.AddWebReference(bstrUrl);
+            get
+            {
+                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
+                var value = (IEnumValue)_projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralBrowseObjectProperties.OutputTypeEx.GetValueAsync);
+                return (prjOutputTypeEx)Convert.ToInt32(value.DisplayName);
+            }
+
+            set
+            {
+                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
+                RunFuncTaskSynchronously(configurationGeneralBrowseObjectProperties.OutputTypeEx.SetValueAsync, (object)value);
+            }
         }
-
-        public void Refresh()
-        {
-            _vsProject.Refresh();
-        }
-
-        public void CopyProject(string bstrDestFolder, string bstrDestUNCPath, prjCopyProjectOption copyProjectOption, string bstrUsername, string bstrPassword)
-        {
-            _vsProject.CopyProject(bstrDestFolder, bstrDestUNCPath, copyProjectOption, bstrUsername, bstrPassword);
-        }
-
-        public void Exec(prjExecCommand command, int bSuppressUI, object varIn, out object pVarOut)
-        {
-            _vsProject.Exec(command, bSuppressUI, varIn, out pVarOut);
-        }
-
-        public void GenerateKeyPairFiles(string strPublicPrivateFile, string strPublicOnlyFile = "0")
-        {
-            _vsProject.GenerateKeyPairFiles(strPublicPrivateFile, strPublicOnlyFile);
-        }
-
-        public string GetUniqueFilename(object pDispatch, string bstrRoot, string bstrDesiredExt)
-        {
-            return _vsProject.GetUniqueFilename(pDispatch, bstrRoot, bstrDesiredExt);
-        }
-
-        public VSLangProj.References References => _vsProject.References;
-
-        public BuildManager BuildManager => _vsProject.BuildManager;
-
-        public DTE DTE => _vsProject.DTE;
-
-        public Project Project => _vsProject.Project;
-
-        public ProjectItem WebReferencesFolder => _vsProject.WebReferencesFolder;
-
-        public string TemplatePath => _vsProject.TemplatePath;
-
-        public bool WorkOffline { get => _vsProject.WorkOffline; set => _vsProject.WorkOffline = value; }
-
-        public Imports Imports => _vsProject.Imports;
-
-        public VSProjectEvents Events => _vsProject.Events;
 
         // Implementation of VsLangProj.ProjectProperties
-        public string __id => throw new NotImplementedException();
-
-        public object __project => throw new NotImplementedException();
-
-        public string StartupObject { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public prjOutputType OutputType
         {
             get
@@ -102,7 +69,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
             }
         }
 
-        public string RootNamespace { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public string AssemblyName
         {
             get
@@ -117,9 +83,41 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
                 RunFuncTaskSynchronously(configurationGeneralProperties.AssemblyName.SetValueAsync, (object)value);
             }
         }
+
+        public string FullPath
+        {
+            get
+            {
+                var configurationGeneralProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralPropertiesAsync);
+                return _projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralProperties.TargetPath.GetEvaluatedValueAtEndAsync);
+            }
+        }
+
+        public string ExtenderCATID => null;
+
+        public string AbsoluteProjectDirectory
+        {
+            get
+            {
+                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
+                return _projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralBrowseObjectProperties.FullPath.GetEvaluatedValueAtEndAsync);
+            }
+        }
+
+        public string __id => throw new NotImplementedException();
+
+        public object __project => throw new NotImplementedException();
+
+        public string StartupObject { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
+        public string RootNamespace { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public string AssemblyOriginatorKeyFile { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public string AssemblyKeyContainerName { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public prjOriginatorKeyMode AssemblyOriginatorKeyMode { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public bool DelaySign { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public string WebServer => throw new NotImplementedException();
@@ -141,18 +139,12 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
         public prjWebAccessMethod ActiveWebAccessMethod => throw new NotImplementedException();
 
         public prjScriptLanguage DefaultClientScript { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public prjTargetSchema DefaultTargetSchema { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public prjHTMLPageLayout DefaultHTMLPageLayout { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public string FileName { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
-        public string FullPath
-        {
-            get
-            {
-                var configurationGeneralProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralPropertiesAsync);
-                return _projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralProperties.TargetPath.GetEvaluatedValueAtEndAsync);
-            }
-        }
+        public prjTargetSchema DefaultTargetSchema { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
+        public prjHTMLPageLayout DefaultHTMLPageLayout { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
+        public string FileName { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public string LocalPath => throw new NotImplementedException();
 
@@ -167,53 +159,20 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
 
         public object ExtenderNames => throw new NotImplementedException();
 
-        public string ExtenderCATID => null;
-
         public string ApplicationIcon { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public prjOptionStrict OptionStrict { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public string ReferencePath { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public string OutputFileName => throw new NotImplementedException();
 
-        public string AbsoluteProjectDirectory
-        {
-            get
-            {
-                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
-                return _projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralBrowseObjectProperties.FullPath.GetEvaluatedValueAtEndAsync);
-            }
-        }
-
         public prjOptionExplicit OptionExplicit { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
         public prjCompare OptionCompare { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public prjProjectType ProjectType => throw new NotImplementedException();
 
         public string DefaultNamespace { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-        public prjOutputTypeEx OutputTypeEx
-        {
-            get
-            {
-                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
-                var value = (IEnumValue)_projectVsServices.ThreadingService.JoinableTaskFactory.Run(configurationGeneralBrowseObjectProperties.OutputTypeEx.GetValueAsync);
-                return (prjOutputTypeEx)Convert.ToInt32(value.DisplayName);
-            }
-
-            set
-            {
-                var configurationGeneralBrowseObjectProperties = _projectVsServices.ThreadingService.JoinableTaskFactory.Run(_projectProperties.Value.GetConfigurationGeneralBrowseObjectPropertiesAsync);
-                RunFuncTaskSynchronously(configurationGeneralBrowseObjectProperties.OutputTypeEx.SetValueAsync, (object)value);
-            }
-        }
-
-        private void RunFuncTaskSynchronously<T>(Func<T, Task> asyncActionT, T value)
-        {
-            _projectVsServices.ThreadingService.JoinableTaskFactory.Run(
-                async () =>
-                {
-                    await asyncActionT(value).ConfigureAwait(false);
-                });
-        }
     }
 }
